backend/requirements.txt
# ─────────────────────────────────────────────────────────────────────────────
# Arokah — Development Dependencies
# For production, see requirements_prod.txt
# ─────────────────────────────────────────────────────────────────────────────

# Core
django>=6.0,<7.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
drf-spectacular>=0.27,<1.0

# Environment
python-dotenv>=1.0,<2.0

# Images
Pillow>=10.0,<12.0

# HTTP (weather service)
requests>=2.31,<3.0

# Calendar integrations
google-auth-oauthlib>=1.2,<2.0
google-auth-httplib2>=0.2,<1.0
google-api-python-client>=2.130,<3.0
msal>=1.28,<2.0
caldav>=1.3,<2.0
icalendar>=5.0,<6.0
cryptography>=42.0,<44.0
python-dateutil>=2.9,<3.0
pytz>=2024.1

# Database (dev uses SQLite built-in; prod needs psycopg2-binary)
# psycopg2-binary>=2.9,<3.0

# AI — uncomment to activate live agents
mistralai>=1.0,<2.0

# ── Dev & Testing ─────────────────────────────────────────────────────────────
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0
pytest-cov>=5.0,<6.0
factory-boy>=3.3,<4.0

# Task queue (dev)
celery>=5.3,<6.0
django-celery-beat>=2.7,<3.0

# WebSockets / real-time (Django Channels)
channels>=4.0,<5.0
channels-redis>=4.2,<5.0
daphne>=4.1,<5.0
